move alignment related function to alignment.py
