+++ onmt/utils/__init__.py
@@ -1,6 +1,6 @@
 """Module defining various utilities."""
-from onmt.utils.misc import split_corpus, aeq, use_gpu, set_random_seed, \
-    make_batch_align_matrix
+from onmt.utils.misc import split_corpus, aeq, use_gpu, set_random_seed
+from onmt.utils.alignment import make_batch_align_matrix
 from onmt.utils.report_manager import ReportMgr, build_report_manager
 from onmt.utils.statistics import Statistics
 from onmt.utils.optimizers import MultipleOptimizer, \
